meeseeks-api/pyproject.toml
[tool.poetry]
name = "meeseeks-api"
version = "1.0.0"
description = "REST API Engine wrapped around the meeseeks-core"
authors = ["Krishnakanth Alagiri <https://github.com/bearlike>"]
license = "MIT"
readme = "README.md"
include = ["*.py"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-restx = "^1.3.0"
python-dotenv = "^1.0.1"
coloredlogs = "^15.0.1"
loguru = "^0.7.2"
langchain-core = "^1.2.8"
langchain-community = "^0.4.1"
langchain-openai = "^1.1.7"
tiktoken = "^0.7.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.0"
autopep8 = "^2.1.0"
pylint = "^3.1.0"
ruff = "^0.6.9"
mypy = "^1.11.2"
pytest-cov = "^7.0.0"

[tool.mypy]
files = [
  "__init__.py",
  "api/**/*.py",
  "backend.py",
]
explicit_package_bases = true
mypy_path = ["."]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
